+++ modules/groovy/ui/src/sap/commerce/toolset/groovy/actionSystem/GroovySpringContextModeActionGroup.kt
@@ -1,0 +1,51 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package sap.commerce.toolset.groovy.actionSystem
+
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.actionSystem.DefaultActionGroup
+import com.intellij.openapi.actionSystem.ex.ActionUtil
+import sap.commerce.toolset.groovy.GroovyConstants
+import sap.commerce.toolset.groovy.SpringContextMode
+import sap.commerce.toolset.i18n
+import sap.commerce.toolset.ui.ActionButtonWithTextAndDescription
+
+class GroovySpringContextModeActionGroup : DefaultActionGroup() {
+
+    init {
+        templatePresentation.putClientProperty(ActionUtil.SHOW_TEXT_IN_TOOLBAR, true)
+        templatePresentation.putClientProperty(ActionUtil.COMPONENT_PROVIDER, ActionButtonWithTextAndDescription(this))
+    }
+
+    override fun getActionUpdateThread() = ActionUpdateThread.BGT
+
+    override fun update(e: AnActionEvent) {
+        val vf = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
+        val currentMode = vf.getUserData(GroovyConstants.KEY_SPRING_CONTEXT_MODE)
+            ?: SpringContextMode.DISABLED
+
+        val mode = i18n("hybris.groovy.actions.springContext.${currentMode.name.lowercase()}")
+
+        e.presentation.icon = currentMode.icon
+        e.presentation.text = i18n("hybris.groovy.actions.springContext.mode", mode)
+    }
+
+}